Changed the folder Location to OneDrive
This happened to be able to collaborate with the laptop also

+++ AlumatTornio/DXF/Arc.cs
@@ -8,13 +8,13 @@ namespace DXF
 {
 	public class Arc
 	{
-		public double CenterX;		//Value 10 in DXF
-		public double CenterY;		//Value 20 in DXF
-		public double Radius;		//Value 40 in DXF
-		public double StartAngle;   //Value 50 in DXF
-		public double EndAngle;		//Value 51 in DXF
+		public float CenterX;		//Value 10 in DXF
+		public float CenterY;		//Value 20 in DXF
+		public float Radius;		//Value 40 in DXF
+		public float StartAngle;   //Value 50 in DXF
+		public float EndAngle;		//Value 51 in DXF
 
-		public Arc(double centerX, double centerY, double radius, double startAngle, double endAngle)
+		public Arc(float centerX, float centerY, float radius, float startAngle, float endAngle)
 		{
 			CenterX = centerX;
 			CenterY = centerY;
